Alterations to Slave Fighter's Freedom
* Fixed furniture being placed on literal nothing. T_null is bad for
you, terrain definition in mapgen that has furniture placed on it should
be the same as the floor around it, shouldn't be the embodiment of
nonexistance. :V
* Fixed nameless NPCs caused by using the old "name+" instead of
"name_suffix"
* Gave named NPCs "unique_name" like the arsonist uses.
* Added a variant of NPC and faction for friendly slave fighters, acting
like the captives in More Locations' bandit outpost. Basically turns the
sketchy gladiator house from Gangbang Simulator 2017 to a mob of NPCs
fighting in two teams.