Changes to labs/command center
* Changed around the layout of the Unknown Lab, giving some walls and
moving items/monsters so they don't try to spawn in walls. Less of an
instant gang-bang. Climax of it is a formerly-sealed off room with
scattered metal wreckage, containg you-know-who.
* Reworked the makeshift command center so it only has to use one
adjacent field, the one with the RV parked in it. Gave it a fence too
and a different overmap terrain name/symbol to not look so weird on the
map.
* Mucked with the bio-weapon lab. Still has failed bio weapon fields
around it, but once again the place with the vehicle spawned is remade
to look like a fenced-in dirt lot. This one's got smashed up fences in
spots, to look fittingly fucked up. o3o